+++ source/refined-github.ts
@@ -150,7 +150,7 @@ import './features/quick-review.js';
 import './features/pr-jump-to-first-non-viewed-file.js';
 import './features/keyboard-navigation.js';
 import './features/vertical-front-matter.js';
-import './features/use-first-commit-message-for-new-prs.js';
+import './features/pr-first-commit-title.js';
 import './features/linkify-user-edit-history-popup.js';
 import './features/clean-repo-filelist-actions.js';
 import './features/prevent-duplicate-pr-submission.js';
